requirements.txt
uvicorn
numpy
fastapi
pillow
torch
torchvision
onnx
onnxruntime
matplotlib
requests
opencv-python
beautifulsoup4